// application_tracker/src/app/components/models/submission.ts
export interface Submission {
    submissionDate: Date;
    vendorCompany: string;
    emailId: string;
    phoneNumber: string;
    implementationClient: string;
    clientName: string;
    payRate: number;
    isSubmitted: boolean;
    typeOwnRecruiter: string;
    submissionStatus: string;
    leadName: string;


}